wolfDemo: pixel_t -> uint16_t RGB565 framebuffer
The wolfDemo STM32U585 port now renders directly into a 240x200 RGB565
buffer (no intermediate 8 bpp framebuffer + conversion pass). Engine
column drawers, V_DrawPatchList, and the splash/finale paths in the
glue write uint16_t pixels through a precomputed lit_lut.

- doomtype.h: pixel_t becomes uint16_t.
- r_defs.h: lighttable_t pinned to uint8_t (colormaps stay
  palette-indexed; lit_lut composes them into RGB565 at write time).
- v_video.c: V_DrawPatchList run-decoders rewritten to track an
  explicit x_cursor in 320-space. Per pixel: drop x where (x & 3)==3
  and write palette_rgb565[pal[v]] at v_xsq(x) on a DISP_W=240 row
  stride. Run-end break clamps xc >= xc_end (some vanilla patches
  emit a final run that slightly overshoots the patch width; the
  upstream byte writer absorbed it via in-row OOB writes that are
  corrupt with halfword-stride RGB565).

+++ src/doomtype.h
@@ -119,7 +119,11 @@ typedef uint8_t boolean;
 #endif
 
 typedef uint8_t byte;
-typedef uint8_t pixel_t;
+/* wolfDemo STM32 port: framebuffer is RGB565. Column drawers write
+ * uint16_t pixels via a precomputed lit_lut to scan_full_buf and the
+ * panel DMA picks them up in MSB-first byte order. lighttable_t stays
+ * uint8_t (palette-indexed colormap) - see r_defs.h. */
+typedef uint16_t pixel_t;
 typedef int16_t dpixel_t;
 
 #if !DOOM_SMALL
